# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a documentation of all endpoints in the API.",
        "title": "Tech Challenge API",
        "contact": {},
        "version": "3.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/liveness": {
            "get": {
                "description": "get the status of http.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Show the status of http.",
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pagamento/checkout/{pedidoId}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pagamento"
                ],
                "summary": "checkout de um pagamento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id do pedido a ser feito o checkout",
                        "name": "pedidoId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "status permitido: aprovado | recusado",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Pagamento"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/commons.MessageResponse"
                        }
                    }
                }
            }
        },
        "/pagamento/{pedido_id}": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pagamento"
                ],
                "summary": "pega um pagamento por pedido id",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id do pedido",
                        "name": "pedido_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Pagamento"
                        }
                    }
                }
            }
        },
        "/readiness": {
            "get": {
                "description": "get the status of http.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Show the status of http.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "commons.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Pagamento": {
            "type": "object",
            "required": [
                "pedido_id",
                "tipo"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pedido_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                },
                "update": {
                    "type": "string"
                },
                "valor": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}